move header with search bar into own component and fix bug with invalid classIdRange filters when filters are cleared
